+++ src/app/(with-header)/layout.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 export default function Layout({ children }: { children: React.ReactNode }) {
   return <>{children}</>;
 }
